+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Fixed bug in ScenarioDbManager._read_scenario_tables_from_db
 - Fixed bug in ScenarioDbManager.update_scenario_output_tables_in_db
+- Fixed bug in ScenarioManager.add_file_as_data_asset (when using CPDaaS)
 ### Added
 - Added DataManager.set_parameters()
 - Added print of num rows and columns inserting in ScenarioDbManager.update_scenario_output_tables_in_db